requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==2.3.3
Flask-WTF==1.2.2
google-api-core==2.25.0rc1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
gspread==5.12.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1